Putting `get_chunks` in `utils.py`
